+++ src/mml_utils/scripts/extract_mml_output.py
@@ -20,6 +20,11 @@ from loguru import logger
 from mml_utils.parse.parser import extract_mml_data
 from mml_utils.parse.target_cuis import TargetCuis
 
+try:
+    import pandas as pd
+except ImportError:
+    pd = None
+
 MML_FIELDNAMES = [
     'event_id', 'docid', 'filename', 'matchedtext', 'conceptstring', 'cui', 'preferredname', 'start', 'length',
 ]
@@ -90,6 +95,7 @@ def extract_mml(note_directories: List[pathlib.Path], outdir: pathlib.Path, cui_
     outdir.mkdir(exist_ok=True)
     note_outfile = outdir / f'notes_{now}.csv'
     mml_outfile = outdir / f'mml_{now}.csv'
+    cuis_by_doc_outfile = outdir / f'cuis_by_doc.csv'
 
     if add_fieldname:
         global MML_FIELDNAMES
@@ -104,6 +110,7 @@ def extract_mml(note_directories: List[pathlib.Path], outdir: pathlib.Path, cui_
     build_extracted_file(note_directories, target_cuis, note_outfile, mml_outfile,
                          output_format, encoding, exclude_negated, output_directories=output_directories,
                          mm_encoding=mm_encoding)
+    build_pivot_table(mml_outfile, cuis_by_doc_outfile)
     return note_outfile, mml_outfile
 
 
@@ -202,6 +209,19 @@ def build_extracted_file(note_directories, target_cuis, note_outfile, mml_outfil
     if missing_note_dict:
         logger.warning(f'''All Missing Note Dict: '{"','".join(missing_note_dict)}' ''')
     logger.info(f'Completed successfully.')
+
+
+def build_pivot_table(mml_file, outfile):
+    if pd is None:
+        logger.warning(f'Unable to build pivot table: please install pandas `pip install pandas` and try again.')
+        return
+    df = pd.read_csv(mml_file, usecols=['docid', 'cui'])
+    n_cuis = df['cui'].nunique()
+    n_docs = df['docid'].nunique()
+    df['count'] = 1
+    df = df.pivot_table(index='docid', columns='cui', values='count', fill_value=0, aggfunc=sum).reset_index()
+    df.to_csv(outfile, index=False)
+    logger.info(f'Output {n_cuis} CUIs for {n_docs} documents to: {outfile}.')
 
 
 def extract_data(note_directories: List[pathlib.Path], *, target_cuis=None, encoding='utf8', mm_encoding='cp1252',