+++ src/mml_utils/parse/json.py
@@ -4,7 +4,7 @@ import pathlib
 from mml_utils.parse.target_cuis import TargetCuis
 
 
-def extract_mml_from_json_data(data, filename, *, target_cuis: TargetCuis=None, extras=None):
+def extract_mml_from_json_data(data, filename, *, target_cuis: TargetCuis = None, extras=None):
     """
 
     :param data:
@@ -15,32 +15,34 @@ def extract_mml_from_json_data(data, filename, *, target_cuis: TargetCuis=None, 
     """
     filename = pathlib.Path(filename)
     i = 0
+    if target_cuis is None:
+        target_cuis = TargetCuis()
     for el in data:
         for event in el['evlist']:
             for cui in target_cuis.get_target_cuis(event['conceptinfo']['cui']):
                 semtype = event['conceptinfo']['semantictypes'][0] if event['conceptinfo']['semantictypes'] else ''
                 data = {**{
-                           'event_id': f'{filename.stem}_{i}',
-                           'filename': filename,
-                           'docid': filename.stem,
-                           'matchedtext': event['matchedtext'],
-                           'conceptstring': event['conceptinfo']['conceptstring'],
-                           'cui': cui,
-                           'preferredname': event['conceptinfo']['preferredname'],
-                           'start': event['start'],
-                           'length': event['length'],
-                           'end': event['start'] + event['length'],
-                           'evid': event['id'],
-                           'negated': el.get('negated', None),
-                           'semantictype': semtype,
-                           'source': event['conceptinfo']['sources'][0],
-                           'all_sources': ','.join(event['conceptinfo']['sources']),
-                           'all_semantictypes': ','.join(event['conceptinfo']['semantictypes']),
-                       }, **{
-                           s: 1 for s in event['conceptinfo']['sources']
-                       }, **{
-                           s: 1 for s in event['conceptinfo']['semantictypes']
-                       }}
+                    'event_id': f'{filename.stem}_{i}',
+                    'filename': filename,
+                    'docid': filename.stem,
+                    'matchedtext': event['matchedtext'],
+                    'conceptstring': event['conceptinfo']['conceptstring'],
+                    'cui': cui,
+                    'preferredname': event['conceptinfo']['preferredname'],
+                    'start': event['start'],
+                    'length': event['length'],
+                    'end': event['start'] + event['length'],
+                    'evid': event['id'],
+                    'negated': el.get('negated', None),
+                    'semantictype': semtype,
+                    'source': event['conceptinfo']['sources'][0],
+                    'all_sources': ','.join(event['conceptinfo']['sources']),
+                    'all_semantictypes': ','.join(event['conceptinfo']['semantictypes']),
+                }, **{
+                    s: 1 for s in event['conceptinfo']['sources']
+                }, **{
+                    s: 1 for s in event['conceptinfo']['semantictypes']
+                }}
                 if extras:
                     data |= extras
                 yield data