+++ src/mml_utils/__init__.py
@@ -1,2 +1,2 @@
 """Command line utilities to support the use/deployment of Metamaplite and cTAKES."""
-__version__ = '0.3.6'
+__version__ = '0.4.0'