+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Types of changes:
 
 * `pandas` as required rather than optional
 * Generate pivot table when running `extract_mml` to show CUI counts note
+* Documentation on using Metamap
 
 ### Fixed